+++ src/main/java/org/leisureup/global/auth/social/GoogleOAuthClient.java
@@ -2,6 +2,7 @@ package org.leisureup.global.auth.social;
 
 import lombok.*;
 import lombok.extern.slf4j.*;
+import org.leisureup.global.auth.dto.api.*;
 import org.leisureup.global.auth.dto.request.SignInUpRequest.*;
 import org.leisureup.global.auth.social.internal.*;
 import org.springframework.cloud.openfeign.*;
